requirements.txt
Django==2.3.19
python-decouple
django_compressor
django-npm
flake8
coverage
python-coveralls
django-libsass
django-compressor-postcss
django-extensions
djangorestframework
django-filter
django-compressor-toolkit
django-js-reverse
django-colorful
django-constance
django-picklefield
requests
psycopg2
gunicorn
whitenoise
roman
nltk
numpy
scikit-learn
scipy
kneed
beautifulsoup4
django-allauth
celery
django-import_export
django-celery-beat
django-celery-results
